-Changed species opinion storage to floats, fixing issue with the initial value being passed as a double which was throwing an exception when extracted as a float. -grooming -Initial implementation of species-species opinion display in pedia.

+++ universe/Effect.cpp
@@ -1223,8 +1223,8 @@ void SetSpeciesSpeciesOpinion::Execute(const ScriptingContext& context) const {
     if (rated_species_name.empty())
         return;
 
-    double initial_opinion = GetSpeciesManager().SpeciesSpeciesOpinion(opinionated_species_name, rated_species_name);
-    double opinion = m_opinion->Eval(ScriptingContext(context, initial_opinion));
+    float initial_opinion = GetSpeciesManager().SpeciesSpeciesOpinion(opinionated_species_name, rated_species_name);
+    float opinion = m_opinion->Eval(ScriptingContext(context, initial_opinion));
 
     GetSpeciesManager().SetSpeciesSpeciesOpinion(opinionated_species_name, rated_species_name, opinion);
 }
